requirements.txt
requests>=2.25.0
behave>=1.2.6
PyYAML>=5.3.1
jsonschema>=3.2.0
setuptools>=51.0.0